+++ CHANGELOG.md
@@ -5,6 +5,11 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ## Unreleased
 
+### Added
+
+* Added support for mapping `types::Timestamp` to/from `chrono::NaiveDateTime`.
+  Add `features = ["chrono"]` to enable.
+
 ### Changed
 
 * Moved most of our top level trait exports into a prelude module, and